# openapi/mm-agent-worker.v1.yaml
openapi: 3.0.3
info:
  title: Magnus Flipper v1 – MM-Agent ↔ Worker API
  version: 1.0.0
  description: >
    DB-lite ingestion contract between the Vercel-hosted mm-agent UI
    and the worker ingestion service. Stateless, no DB writes.

servers:
  - url: https://worker.flipperagents.com

security:
  - MmAgentAuth: []

components:
  securitySchemes:
    MmAgentAuth:
      type: apiKey
      in: header
      name: x-mm-agent-token

  schemas:
    IngestRunRequest:
      type: object
      required:
        - requestId
        - initiatedBy
        - mode
        - marketplaces
        - searches
      properties:
        requestId:
          type: string
          format: uuid
        initiatedBy:
          type: string
          enum: [mm-agent]
        mode:
          type: string
          enum: [db-lite]
        marketplaces:
          type: array
          items:
            type: string
            enum: [facebook, gumtree, vinted]
        searches:
          type: array
          items:
            $ref: "#/components/schemas/SearchSpec"

    SearchSpec:
      type: object
      required:
        - searchId
        - marketplace
        - query
        - location
      properties:
        searchId:
          type: string
        marketplace:
          type: string
          enum: [facebook, gumtree, vinted]
        query:
          type: string
        location:
          type: string
          enum: [US, UK]
        filters:
          type: object
          additionalProperties: true

    IngestAcceptedResponse:
      type: object
      properties:
        requestId:
          type: string
        status:
          type: string
          enum: [accepted]
        startedAt:
          type: string
          format: date-time
        estimatedDurationSec:
          type: integer

    IngestStatusResponse:
      type: object
      properties:
        requestId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, partial, failed]
        progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            failed:
              type: integer
        startedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ListingItem:
      type: object
      properties:
        title:
          type: string
        price:
          type: number
        currency:
          type: string
        location:
          type: string
        url:
          type: string
        images:
          type: array
          items:
            type: string
        sellerName:
          type: string
        postedAt:
          type: string
          format: date-time

    IngestResult:
      type: object
      properties:
        marketplace:
          type: string
        searchId:
          type: string
        query:
          type: string
        location:
          type: string
        listingsFound:
          type: integer
        durationMs:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListingItem"

    IngestResultsResponse:
      type: object
      properties:
        requestId:
          type: string
        mode:
          type: string
        completedAt:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/IngestResult"

    HealthResponse:
      type: object
      properties:
        status:
          type: string
        ingestionEnabled:
          type: boolean
        mode:
          type: string
        uptimeSec:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

paths:
  /ingest/run:
    post:
      summary: Trigger ingestion run
      security:
        - MmAgentAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRunRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestAcceptedResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /ingest/status/{requestId}:
    get:
      summary: Poll ingestion status
      security:
        - MmAgentAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestStatusResponse"

  /ingest/results/{requestId}:
    get:
      summary: Fetch ingestion results
      security:
        - MmAgentAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResultsResponse"

  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"